Initial fix for Issue #46 moving updateSSComponent out of SSCombos.
